requirements.txt
infi.systray
SoundCard
numpy
pywin32
comtypes